Move state list widget into its own file
