+++ app/lib/screens/dashboard_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import '../services/firebase_service.dart';
+import '../widgets/state_list.dart';
 
 class DashboardPage extends StatefulWidget {
   const DashboardPage({super.key});
@@ -72,43 +73,10 @@ class _DashboardPageState extends State<DashboardPage> {
           ),
           ConstrainedBox(
               constraints: const BoxConstraints(maxWidth: 400),
-              child: YourWidget(
+              child: StateList(
                   states: states, handleSwitchClick: handleSwitchClick))
         ],
       ),
-    );
-  }
-}
-
-class YourWidget extends StatelessWidget {
-  final List<Map<String, dynamic>> states;
-  final Function(Map<String, dynamic>) handleSwitchClick;
-
-  const YourWidget(
-      {super.key, required this.states, required this.handleSwitchClick});
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView.builder(
-      shrinkWrap: true,
-      itemCount: states.length,
-      itemBuilder: (context, index) {
-        final state = states[index];
-        return ListTile(
-          leading: Icon(
-            state['icon'],
-            color: state['state'] == 1 ? Colors.green : null,
-          ),
-          title: Text(
-              '${state['name']} State: ${state['state'] == -1 ? '...' : state['state'] == 1 ? 'On' : 'Off'}'),
-          trailing: Switch(
-            value: state['state'] == 1,
-            onChanged: (bool value) {
-              handleSwitchClick(state);
-            },
-          ),
-        );
-      },
     );
   }
 }
